action parameter: I added an action parameter in the constructor, which allows you to pass a function that should be executed when the button is clicked. For example, you could pass a function to "like" a post or "next post."
draw() method: This method is used to draw the button on the screen. It uses pygame.draw.rect() to draw the button with the given dimensions and color.

button_in_mouse() method: This checks if the mouse position is within the button’s rectangle. It returns True if the mouse click is inside the button's area.

click_action() method: This method checks if the mouse click is inside the button using button_in_mouse(). If the button is clicked and the action is defined, it triggers the action.